// types/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from "contentful";
import { Document } from "@contentful/rich-text-types";

export interface IAboutMeFields {
  sort: number;
  title: string;
  description: Document;
}

export interface IAboutMe extends Entry<IAboutMeFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "aboutMe";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IAboutMePictureFields {
  title: string;
  photo: Asset;
}

export interface IAboutMePicture extends Entry<IAboutMePictureFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "aboutMePicture";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IContactsFields {
  sort: number;
  title: string;
  description: Document;
}

export interface IContacts extends Entry<IContactsFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "contacts";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IProjectsFields {
  sort: number;
  title: string;
  description: string;
  technologies: string[];
  source: string;
  demo: string;
  photo: Asset;
  photoAlt: string;
}

export interface IProjects extends Entry<IProjectsFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "projects";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ISocialLinksFields {
  sort: number;
  title: string;
  buttonTitle: string;
  url: string;
  icon: Asset;
}

export type TSocialLink = {
  id: string;
  sort: number;
  title: string;
  buttonTitle: string;
  url: string;
  icon: string;
};

export interface ISocialLinks extends Entry<ISocialLinksFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "socialLinks";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export type CONTENT_TYPE = "aboutMe" | "aboutMePicture" | "contacts" | "projects" | "socialLinks";

export type LOCALE_CODE = "en-US" | "pl";

export type CONTENTFUL_DEFAULT_LOCALE_CODE = "en-US";